Alternate approach
I started this to remove the first "of" and change the "it" that refers to the athlete to "he" since people are not "its". Then, I tried to follow the Given, do, then output pattern in this suggestion, and then I just thought it read better if "you" were the athlete.  And suddenly the whole thing had been rewritten. What do you think?