Simplified the initial example for if & if/else.
I think the initial example is too difficult (and too far removed from everyday experience) for beginners, especially non-mathematicians. As much as possible learning should be based in familiar contexts, so the learner can focus only on the new information.

I am very happy for my example to be altered / improved, it's just a suggestion.